家計簿登録のテスト(異常系)を追加 / テストクラス名の変更
